Fix pull request #6
Due to architecture of restkit, the default value
of argument `headers` is ignored.

Methods `get`, `post`, `head`, `put` and `delete`
call method `request` with argument `headers` set to `None`
if `headers` was unset when the parent method was called.

Thus, to set a default value for `headers` in `request`,
we need to be sure that it is not already set to `None`.